+++ packages/mcp/pyproject.toml
@@ -8,7 +8,7 @@ license = { text = "MIT" }
 authors = [{ name = "Keycard", email = "support@keycard.ai" }]
 dependencies = [
     "keycardai-oauth>=0.9.0",
-    "keycardai-starlette-oauth>=0.1.0",
+    "keycardai-starlette>=0.1.0",
     "mcp>=1.13.1",
     "pydantic>=2.11.7",
     "httpx>=0.27.2",
