+++ packages/mcp/src/keycardai/mcp/server/handlers/jwks.py
@@ -1,9 +1,9 @@
 """JWKS endpoint handler.
 
-Re-exported from keycardai.starlette_oauth.handlers.jwks for backward compatibility.
-Canonical import: ``from keycardai.starlette_oauth.handlers.jwks import jwks_endpoint``
+Re-exported from keycardai.starlette.handlers.jwks for backward compatibility.
+Canonical import: ``from keycardai.starlette.handlers.jwks import jwks_endpoint``
 """
 
-from keycardai.starlette_oauth.handlers.jwks import jwks_endpoint
+from keycardai.starlette.handlers.jwks import jwks_endpoint
 
 __all__ = ["jwks_endpoint"]
