+++ packages/mcp/src/keycardai/mcp/server/middleware/bearer.py
@@ -1,10 +1,10 @@
 """Bearer token authentication middleware.
 
-Re-exported from keycardai.starlette_oauth.middleware.bearer for backward compatibility.
-Canonical import: ``from keycardai.starlette_oauth.middleware import BearerAuthMiddleware``
+Re-exported from keycardai.starlette.middleware.bearer for backward compatibility.
+Canonical import: ``from keycardai.starlette.middleware import BearerAuthMiddleware``
 """
 
-from keycardai.starlette_oauth.middleware.bearer import (
+from keycardai.starlette.middleware.bearer import (
     BearerAuthMiddleware,
     _get_bearer_token,
     _get_oauth_protected_resource_url,
