+++ packages/mcp/src/keycardai/mcp/server/shared/starlette.py
@@ -1,9 +1,9 @@
 """Starlette request helpers.
 
-Re-exported from keycardai.starlette_oauth.shared.starlette for backward compatibility.
-Canonical import: ``from keycardai.starlette_oauth.shared import get_base_url``
+Re-exported from keycardai.starlette.shared.starlette for backward compatibility.
+Canonical import: ``from keycardai.starlette.shared import get_base_url``
 """
 
-from keycardai.starlette_oauth.shared.starlette import SUPPORTED_PROTOCOLS, get_base_url
+from keycardai.starlette.shared.starlette import SUPPORTED_PROTOCOLS, get_base_url
 
 __all__ = ["SUPPORTED_PROTOCOLS", "get_base_url"]
