+++ packages/starlette/src/keycardai/starlette/shared/starlette.py
@@ -1,6 +1,7 @@
 """Starlette request helpers for proxy-aware URL construction."""
 
 from pydantic import AnyHttpUrl
+
 from starlette.requests import Request
 
 SUPPORTED_PROTOCOLS = ["http", "https"]
