+++ pyproject.toml
@@ -60,7 +60,7 @@ exclude = []
 [tool.uv.sources]
 # Workspace sources will be defined here as packages are added
 keycardai-oauth = { workspace = true }
-keycardai-starlette-oauth = { workspace = true }
+keycardai-starlette = { workspace = true }
 keycardai-mcp-fastmcp = { workspace = true }
 keycardai-mcp = { workspace = true }
